fix: resolve all clippy warnings for clean CI
- Replace approximate PI value in test fixture
- Convert loop-variable-only-used-to-index patterns to iterators
- Change &mut Vec to &mut [_] in function signatures
- Remove unnecessary u64 casts
- Replace redundant closures with direct function references
- Allow too_many_arguments on test helpers
- Prefix unused assignments with underscore

+++ trustchain-core/src/protocol.rs
@@ -993,14 +993,14 @@ mod tests {
                 .create_proposal(
                     &bob.pubkey(),
                     serde_json::json!({"i": i}),
-                    Some(1000 + i as u64),
+                    Some(1000 + i),
                 )
                 .unwrap();
             assert_eq!(proposal.sequence_number, i);
 
             bob.receive_proposal(&proposal).unwrap();
             let agreement = bob
-                .create_agreement(&proposal, Some(1001 + i as u64))
+                .create_agreement(&proposal, Some(1001 + i))
                 .unwrap();
             assert_eq!(agreement.sequence_number, i);
